% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abnorm_rm.R
\name{abnorm_rm}
\alias{abnorm_rm}
\title{Remove Abnormal Values}
\usage{
abnorm_rm(var, data, min, max)
}
\arguments{
\item{var}{A numeric, continuous variable given in df$var format.}

\item{data}{Source data frame}

\item{min}{The lower threshold of expected values.}

\item{max}{The upper threshold of expected values.}
}
\value{
A data frame.
}
\description{
Remove Abnormal Values
}
