% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox_lam.R
\name{boxcox_lam}
\alias{boxcox_lam}
\title{Box-Cox Lambda Overview}
\usage{
boxcox_lam(var, formula, data)
}
\arguments{
\item{var}{A continuous variable.}

\item{formula}{a regression formula, raw text or stored as an object, using lm().}

\item{data}{Original dataframe.}
}
\value{
A plot and table of lambda values.
}
\description{
Box-Cox Lambda Overview
}
