% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{checkLazyDir}
\alias{checkLazyDir}
\title{Checks and sets for lazyDir}
\usage{
checkLazyDir(lazyDir = NULL, create = TRUE)
}
\arguments{
\item{lazyDir}{Character string of directory to be used for the lazy databases.}

\item{create}{A logical. If \code{TRUE}, it will create the directory specified with \code{lazyDir}}
}
\value{
Directory of lazyDir. If it was not specified, then it returns a
temp directory
}
\description{
Checks and sets for lazyDir
}
\examples{
\dontrun{
checkLazyDir(create=TRUE) # because missing, it will provide a tmpdir
unlink(file.path(tempdir(), "lazyDir"), recursive=TRUE)
}
}
\author{
Eliot McIntire
}

