\name{col.as.row.names}
\alias{col.as.row.names}
\title{Put column values in row names}
\usage{
  col.as.row.names(data, select)
}
\arguments{
  \item{data}{the data frame}

  \item{select}{the column position (can be an integer or a
  character string corresponding to the column name)}
}
\description{
  This function is the inverse of
  \code{\link{row.names.as.col}}
}
\seealso{
  \code{\link{row.names.as.col}}
}

