% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{connect_to_local}
\alias{connect_to_local}
\title{Connect to local database}
\usage{
connect_to_local(name)
}
\arguments{
\item{None.}{}
}
\description{
This connects me to a local MySQL database - you might need to change host and root to work on your computer.
}
\examples{
connect_to_local("my_db")
}
\keyword{MySQL}
\keyword{database}
\keyword{local}

