\name{contains.integer}
\alias{contains.integer}
\title{Is a numeric value an integer?}
\usage{
  contains.integer(x)
}
\arguments{
  \item{x}{the object to test.}
}
\value{
  A boolean
}
\description{
  Test if a numeric object (or integer) contains an
  integer. \code{is.numeric} only tests the class of the
  object such that \code{is.numeric(1)} returns
  \code{FALSE}.
}
\seealso{
  \code{\link{is.integer}}
}

