% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\docType{methods}
\name{copyLazyObjs}
\alias{copyLazyObjs}
\title{Copy individual objects from one lazy database to another}
\usage{
copyLazyObjs(objNames, from = NULL, to = NULL, create = TRUE,
  verbose = FALSE, ...)
}
\arguments{
\item{objNames}{a character vector of object names to copy from one \code{lazyDir} to another}

\item{from}{The source \code{lazyDir}}

\item{to}{The new \code{lazyDir}}

\item{create}{Passed to \code{checkLazyDir}}

\item{verbose}{Logical. If notes about rasters whose file backings were corrected on the
fly from previous, i.e,. wrong, filenames.}

\item{...}{Other arguments passed to lazySave, such as overwrite}
}
\description{
This will copy individual objects from one lazy database to another. This is much slower
than copying a whole lazy database using \code{copyDir}, but it will merge objects from
one database to another.
}
\author{
Eliot McIntire
}

