% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_kable.R
\name{dark_kable}
\alias{dark_kable}
\title{Dark-Theme Friendly Kable}
\usage{
dark_kable(
  x,
  format = "html",
  digits = 2,
  row.names = NA,
  col.names = NA,
  align = NULL,
  caption = NULL,
  label = NULL,
  format.args = list(),
  escape = TRUE,
  decimal.mark = ".",
  table.attr = "style = \\"color:white;\\""
)
}
\arguments{
\item{x}{The data to be printed as a kable, typically a matrix or a dataframe.}

\item{format}{A character string denoting format. This is set to "html" by default.}

\item{digits}{The number of decimals to print, byd efault this is set to 2.}

\item{row.names}{Logical: whether to include row names.}

\item{col.names}{A character vector of column names to be used in the table.}

\item{align}{Alignment of cell contents}

\item{caption}{The caption of the table}

\item{label}{The table reference label}

\item{format.args}{A list of arguments to be passed to format}

\item{escape}{Boolean; whether to escape special characters when producing html or LaTeX tables.}

\item{decimal.mark}{The symbol to be used for decimals, by default this is set to "."}

\item{table.attr}{By default this is set to "style = \"color:white;\"" which together with format = html gives
white text output in RMarkdown.}
}
\value{
A kable-formated table using html styling to achieve white text.
}
\description{
Dark-Theme Friendly Kable
}
\note{
See ?knitr::kable for more details on each argument. In order to get black text on a white background when
knitting, the following CSS code must be added to the start of the RMarkdown file (and NOT in an R code chunk):\if{html}{\out{
<style>
table \{
 background-color: white !important;
 color: black !important;
\}
</style>
}}


This code is also stored in the lazyR package as enable_dark_kable for easy reference.
NB! The code must still be copied in, calling the object will not work.
}
