% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dhist.r
\name{dhist}
\alias{dhist}
\title{dhist function}
\usage{
dhist(data, ...)
}
\arguments{
\item{x}{A vector of discrete data.}
}
\description{
This function allows you create histogram for discrete data.
}
\examples{
x<-c(0,0,0,1,1,2,3,3,4,4,4,4,5,5)
dhist(x)
}
\keyword{dhist}

