% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{freq}
\alias{freq}
\title{Frequency counts}
\usage{
freq(vector, ordered = FALSE)
}
\arguments{
\item{ordered}{If TRUE the returned data frame is returned with most frequent items at top Defaults to FALSE.}
}
\description{
Check the frequencies of values in a vector.
}
\examples{
freq()
}
\keyword{table}

