% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{gg_colour_hue}
\alias{gg_colour_hue}
\title{Re-create ggplot colour palettes}
\usage{
gg_colour_hue(n)
}
\arguments{
\item{None}{}
}
\description{
Get a vector of hexadecimal colours as used by ggplot2 for a given number of groups in a plot.
}
\examples{
gg_colour_hue(3)
}
\keyword{ggplot2}
\keyword{palette}

