% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not In}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{What to search for}

\item{y}{What do search in}
}
\value{
a logical vector of whether values of x are not found in y
}
\description{
Not In
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(1, 2, 3, 4, 6)

x \%!in\% y

}
