\name{is.categorical}
\alias{is.categorical}
\title{Check if the object is categorical (that is not numeric)}
\usage{
  is.categorical(x)
}
\arguments{
  \item{x}{the object to test.}
}
\value{
  A boolean
}
\description{
  Test if an object belongs to class factor, character or
  logical. These three types are defined as categorical.
}
\seealso{
  \code{\link{contains.integer}}
}

