% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{last function}
\usage{
last(v)
}
\arguments{
\item{v}{input a vector.}
}
\description{
This function allows you to call the last element of a vector.
}
\examples{
x<-c(2,5,74,33,50)
last(x)
}
\keyword{last}

