% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_bar.R
\name{lazy_bar}
\alias{lazy_bar}
\title{Draw a Quick and Pretty Bar Chart}
\usage{
lazy_bar(
  xvar,
  yvar,
  data,
  palette = "Dark2",
  fillby = xvar,
  width = 0.9,
  theme = BarTheme,
  pos = "dodge",
  errorpos = ggplot2::position_dodge(),
  xlab = "X Label",
  ylab = "Y Label"
)
}
\arguments{
\item{xvar}{Variable to be displayed on the X axis in df$xvar format.}

\item{yvar}{Variable to be displayed on the Y axis in df$yvar format.}

\item{data}{The data frame containing the data}

\item{palette}{The colour palette to use when drawing the graph.}

\item{fillby}{Optional. the variable to fill by.
By default this is set to xvar.}

\item{width}{The width of each bar, by default this is set to 0.9}

\item{theme}{The theme to use. By default this is set to the build-in
BarTheme theme.}

\item{pos}{Position of the bars, by default this is set to "dodge".
The default should be fine unless you use fill or colour to separate the data.}

\item{errorpos}{Position of the error bars, by default this is set to
position_dodge(). Replace with position_dodge(value) e.g., 0.9}

\item{xlab}{Text to be displayed on the x axis}

\item{ylab}{Text to be displayed on the y axis.}
}
\value{
A bar chart.
}
\description{
Draw a Quick and Pretty Bar Chart
}
\note{
You can augment the function with further ggplot2 calls like
facet_wrap(), labs(), and theme(). If you need to limit
the y-axiz value to zoom in on an effect, add ggplot's
coord_cartesian(ylim = c(min, max)) function. You should also
use ggplot to hide or modify the legend.
}
