% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_contrast_info.R
\name{lazy_contrast_info}
\alias{lazy_contrast_info}
\title{Lazy Contrast Info}
\usage{
lazy_contrast_info(var)
}
\arguments{
\item{var}{The variable to apply contrasts to}

\item{contrasts}{The contrasts to be applied, e.g.,  cbind(c(1, -1)) for simple sum orthogonal contrasts for a factor of two levels.}
}
\value{
contrasts with descriptive outputs.
}
\description{
Lazy Contrast Info
}
