% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_contrasts.R
\name{lazy_contrasts}
\alias{lazy_contrasts}
\title{Lazy Contrasts}
\usage{
lazy_contrasts(var, contrasts)
}
\arguments{
\item{var}{The variable to apply contrasts to}

\item{contrasts}{The contrasts to be applied, e.g.,  cbind(c(1, -1)) for simple sum orthogonal contrasts for a factor of two levels.}
}
\value{
contrasts with descriptive outputs.
}
\description{
Lazy Contrasts
}
