% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_hist.R
\name{lazy_hist}
\alias{lazy_hist}
\title{Easily Create a Pretty Histogram}
\usage{
lazy_hist(
  xvar,
  data,
  theme = HistTheme,
  col = "black",
  fill = "cyan",
  bins = 30,
  normal = TRUE
)
}
\arguments{
\item{xvar}{The variable to be displayed on the x axis in df$xvvar format}

\item{data}{The data frame where the data resides}

\item{theme}{Optional - an object containing a theme.
By default, the built-in HistTheme is used.}

\item{col}{Optional - The outline colour of each bar.
By default this is black.}

\item{fill}{Optional - The fill colour of each bar.
By default this is cyan.}

\item{bins}{Optional - The desired number of bins.
By default this is 30.}

\item{normal}{Toggle whether to display a normal curve overlay.
By default, this is set to TRUE.}
}
\value{
A histogram
}
\description{
Easily Create a Pretty Histogram
}
\note{
This function creates a ggplot2 object. You can further edit this by
using normal ggplot2 syntax. For example, to wrap you could add
lazyR::lazy_hist(xvar, df) + facet_wrap(~cond1 * cond2)
to create a 2x2 histogram grid.
}
