% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_interact.R
\name{lazy_interact}
\alias{lazy_interact}
\title{Lazy Interaction Plots}
\usage{
lazy_interact(
  xvar,
  yvar,
  data,
  groupby,
  colby = groupby,
  shapeby = groupby,
  theme = InteractTheme,
  linesize = 2,
  pointsize = 4,
  position = ggplot2::position_dodge(width = 0.25),
  legendtitle = "Legend Title",
  xlab = "xlabel",
  ylab = "ylabel",
  palette = "Dark2",
  type = 1
)
}
\arguments{
\item{xvar}{This variable will be displayed on the x-axis. Most often this will be categorical (i.e., a condition)}

\item{yvar}{This variable will be displayed on the y-axis. Most often this will be continuous (e.g., RTs)}

\item{data}{The dataframe where the data is stored.}

\item{groupby}{The variable to group by. For type 1 plots (Main effect plots) this is by default set to
match the colour-by variable. Often this needs to be a separate variable for type 2 or 3 plots.}

\item{colby}{The variable to colour by. This should be categorical (i.e., a condition). This can be different from
the grouping variable or the same. In type 2 or 3 plots this should often be a separate condition.}

\item{theme}{The theme stored as an object. By default, the built-in "InteractionTheme" is used.}

\item{linesize}{The size of the main effect lines, the default is set to 2.}

\item{pointsize}{The size of the main effect points, the default is set to 4.}

\item{position}{The position adjustment, by default this is set to position_dodge(width = 0.25) to avoid overlapping.}

\item{legendtitle}{Your desired legend title.}

\item{palette}{The palette to use, by default Dark2 from RColorBrewer is used.}

\item{type}{There are three types of plot available:
1 = Main Effect plot
2 = Individual data points (per Pp) plot
3 = Individual slope (per Pp) plot.}
}
\value{
An interaction plot
}
\description{
Lazy Interaction Plots
}
\note{
This function requires data in a specific format. For type 1 plots the raw data frame will suffice in most cases.

For type 2 plots, an aggregated mean fpr each participant within relevant conditions should be used, e.g.:
Summary <- aggregate(RT ~ Pp  + PhraseType + Switch + Language, data = df, FUN = "mean")
lazyR::lazy_interact(xvar = Summary$PhraseType,
yvar = Summary$RT,
group = Summary$Switch,
col = Summary$Switch,
data = Summary,
type = 2) +
facet_wrap(~ Switch * Language)

Similarly, type 3 plots require an aggregated mean, e.g.:

Summary <- aggregate(RT ~ Pp  + PhraseType + Switch, data = df, FUN = "mean")
lazyR::lazy_interact(xvar = Summary$PhraseType,
yvar = Summary$RT,
group = Summary$Pp,
col = Summary$Switch,
data = Summary,
type = 3) +
facet_wrap(~Switch)

If you get more points or lines than desired when using type 2 or 3 plots, check that the data is in the
correct format - one data point per Pp per condition only.

If the data is in the correct format, remember to apply relevant facet_wraps. In the above example,
the type 2 plot would need facet wrapping by Switch and Language to avoid this problem. In the second example,
the plot would need to be facet wrapped by Switch only. Remember that type 2 plots are more dense than type 1 plots,
and that type 3 plots are in turn more dense than type 2 plots. Faceted plots become increasingly difficult to scale down.

Lastly, a note on grouping. For type 2 and type 3 plots, the group variable should be the participant ID variable,
while the col variable should be a discrete condition grouping variable (see the above examples)
}
