% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_pirate.R
\name{lazy_pirate}
\alias{lazy_pirate}
\title{Generate a Pirate, Box, or Violin Plot}
\usage{
lazy_pirate(
  xvar,
  yvar,
  data,
  fillby = xvar,
  theme = BoxTheme,
  palette = "Dark2",
  dodge = 1,
  xlab = "xlabel",
  ylab = "ylabel",
  type = "Pirate",
  outlier = NA,
  boxwidth = 0.5
)
}
\arguments{
\item{xvar}{The x variable, often a discrete condition.}

\item{yvar}{The y variable, often a continuous variable (e.g., RT)}

\item{data}{The data frame containing the data}

\item{fillby}{Variable to fill by, by default this is set to xvar.}

\item{theme}{The theme to apply, by default this uses the built-in BoxTheme.}

\item{palette}{The palette to fill by, by default this is set to Dark2 from RColorBrewer}

\item{dodge}{Dodge value. By default this is set to 1. Adjust this if box- and/or violin plots do not line up.}

\item{xlab}{Label for the x axis}

\item{ylab}{Label for the y axis}

\item{type}{The type of plot, choose between "Pirate", "Box", and "Violin". The default is set to "Pirate".}

\item{outlier}{Choose whether to display outliers (boxplots only). By default these are hidden (NA)}

\item{boxwidth}{The width of the boxplots, adjust as needed. By default this is set to 0.5.}
}
\value{
A Pirate, box, or violinplot.
}
\description{
Generate a Pirate, Box, or Violin Plot
}
