% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_post_plot.R
\name{lazy_post_plot}
\alias{lazy_post_plot}
\title{Lazy Posterior Distribution Plots}
\usage{
lazy_post_plot(
  data,
  xvar,
  ymin = -1,
  ymax = -2,
  title = NULL,
  HDPI = 0.93,
  colpos = "#2196F3",
  colneg = "#A17C66"
)
}
\arguments{
\item{data}{Required: The dataframe where the data was stored originally - this will typically be a dataframe of samples from the posterior.}

\item{xvar}{Required: The variable or parameter whose posterior is to be plotted. This will often need to be put in `` in the case of interactions}

\item{ymin}{Optional: This will toggle the thickness of the HDPI. Override if the defaults are too thick or thin.}

\item{ymax}{Optional: This also toggles the thickness of the HDPI overlay. Override if the defaults are too thick or thin.}

\item{title}{Optional: The title of the plot. The default is no title.}

\item{HDPI}{Optional: The value of the HDPI. By default set to .93 to differentiate from frequentist confidence intervals.}

\item{colpos}{Optional: The colour of positive areas under the curve. The default is blue.}

\item{colneg}{Optional: The colour of negative areas under the curve. The default is a matte brown.}
}
\value{
A posterior density plot.
}
\description{
Lazy Posterior Distribution Plots
}
