% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_post_ratio.R
\name{lazy_post_ratio}
\alias{lazy_post_ratio}
\title{Lazy Posterior Probability Proportions}
\usage{
lazy_post_ratio(var, data)
}
\arguments{
\item{var}{The variable or parameter to be quantified}

\item{data}{The dataframe containing the variable or parameter. This will most commonly be a dataframe of posterior samples.}
}
\value{
Probability rations above and below zero for the desired parameter.
}
\description{
Lazy Posterior Probability Proportions
}
