% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_qqnorm.R
\name{lazy_qqnorm}
\alias{lazy_qqnorm}
\title{Normal QQ Plot}
\usage{
lazy_qqnorm(
  var,
  data,
  col = "#00BFC4",
  linecol = "Red",
  linesize = 2,
  xlab = "Theoretical Value",
  ylab = "Sampled Value"
)
}
\arguments{
\item{var}{The variable to sample}

\item{data}{The data frame containing the data}

\item{col}{The colour of the qq points, by default this is set to
"#00BFC4" which is a blue colour. Must be in "".}

\item{linecol}{The colour of the normal guide, by default this is set to "Red"
Must be in "".}

\item{linesize}{The width of the normal guide, by default this is set to 2.}

\item{xlab}{The label of the x-axis, by default this is "Theoretical Value".
Must be in "".}

\item{ylab}{The label of the y-axis, by default this is "Sampled Value".
Must be in "".}
}
\value{
A qq plot with a normal distribution guideline.
}
\description{
Normal QQ Plot
}
