% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{lazyExists}
\alias{lazyExists}
\title{Does an object exist in the lazyDir}
\usage{
lazyExists(objName, lazyDir = NULL, exact = TRUE)
}
\arguments{
\item{objName}{A character string indicating the object name}

\item{lazyDir}{the lazyDir to use}

\item{exact}{Should the character string matching be exact, i.e., ^objNames$
in regexp. Passed to \code{lazyLs()}}
}
\value{
Logical
}
\description{
Does an object exist in the lazyDir
}
\examples{
lazyDir(file.path(tempdir(), "lazyDir"), create=TRUE)
a <- rnorm(10)
lazySave(a)
lazyExists("a")
lazyRm("a")
lazyExists("b") # error, does not exist
unlink(file.path(tempdir(), "lazyDir"), recursive=TRUE)
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyLoad2}}
}

