% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{lazyIs}
\alias{lazyIs}
\title{Get object class}
\usage{
lazyIs(objName, class2 = NULL, removeCharacter = TRUE, lazyDir = NULL)
}
\arguments{
\item{objName}{A character string indicating the object name}

\item{class2}{An optional character string of the class to test against}

\item{removeCharacter}{Logical. There is an artifact tag, class:character for any
non-character object. This will be removed automatically in the return, unless this
is FALSE}

\item{lazyDir}{Character string of directory to be used for the lazy databases.}
}
\value{
A character vector equivalent to the an \code{is(objName)} command
}
\description{
Get object class
}
\examples{
lazyDir(file.path(tempdir(), "lazyDir"), create=TRUE)
a <- rnorm(10)
lazySave(a)
lazyIs("a", "numeric")
lazyIs("a") # shows all
lazyRm("a")

\dontrun{
lazyIs("b") # error, does not exist
}

unlink(file.path(tempdir(), "lazyDir"), recursive=TRUE)
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyLoad2}}
}

