% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{lazyLoad2}
\alias{lazyLoad2}
\title{Load lazy objects from a \code{lazyR} database}
\usage{
lazyLoad2(objNames = NULL, md5Hashes = NULL, lazyDir = NULL,
  envir = parent.frame(), verbose = FALSE)
}
\arguments{
\item{objNames}{A character vector of object names to load lazily, usually from a lazyLs call.
If \code{NULL}, all object in lazyDir will be loaded.}

\item{md5Hashes}{Object hash.}

\item{lazyDir}{Character string of directory to be used for the lazy databases.}

\item{envir}{Environment into which the objects are loaded. Default is \code{.GlobalEnv}.}

\item{verbose}{Logical. If notes about rasters whose file backings were corrected on the
fly from previous, i.e,. wrong, filenames.}
}
\value{
Invisibly, the objects read in returned. This function is used for its
side effects, i.e., loading lazy objects.
}
\description{
Load all named objects into the specified environment.
}
\note{
Rasters have a special behaviour. If the filename associated with a raster is
pointing to a file that doesn't exist, then it will try the prepend the \code{lazyDir} to the
filename inside the \code{Raster*} object. The raster filenames can become incorrect if
the absolute path to the file changes due to a changing of operating system or moving
from a removable drive. For example, a removable drive, which was mapped to, say E: in the
computer where the object was originally \code{lazySave}d, gets replugged in but mapped to
F:, then the absolute paths will be wrong. This will be automatically corrected, with a warning,
if the filename can be found by looking in the relative path \code{/rasters/} below
\code{lazyDir}.
}
\examples{
library(SpaDES)
tmpdir <- checkPath(file.path(tempdir(), "lazyDir"), create=TRUE)
obj <- rnorm(10)
# save the obj
lazySave(obj, lazyDir=tmpdir)
# remove the obj
rm(obj)
any(ls()=="obj") # Is FALSE
# load it back in
lazyLoad2("obj", lazyDir=tmpdir)
any(ls()=="obj") # Is TRUE
unlink(tmpdir, recursive = TRUE)
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyRm}}.
}

