% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyLoad.R
\docType{methods}
\name{lazyLoadFromRepo}
\alias{lazyLoadFromRepo}
\title{Lazy load mechanism}
\usage{
lazyLoadFromRepo(artifact, lazyDir = lazyDir(), objName,
  envir = parent.frame(1))
}
\arguments{
\item{artifact}{A character vector of the hash associated with an object in the lazyDir}

\item{lazyDir}{the lazyDir to use}

\item{objName}{Must pass the object name as a character string to assign the lazy evaluation to}

\item{envir}{The environment to assign into (passed to \code{assign.env} arg of \code{delayedAssign})}
}
\value{
An object named objName is assigned into the envir, lazily via \code{delayedAssign}.
}
\description{
This creates the promise to be evaluated later, used internally by \code{lazyLoad2}.
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyLoad2}}
}

