% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{lazyObjectName}
\alias{lazyObjectName}
\title{Get object name with an md5hash}
\usage{
lazyObjectName(md5Hash, lazyDir = NULL)
}
\arguments{
\item{md5Hash}{A character string indicating the md5hash value to look up the object name}

\item{lazyDir}{Character string of directory to be used for the lazy databases.}
}
\value{
An object name
}
\description{
Get object name with an md5hash
}
\examples{
\dontrun{
a <- rnorm(10)
lazyDir(file.path(tempdir(), "lazyDir"), create=TRUE)
lazySave(a) # default uses an tempdir() call for location
hash <- lazyLs("a", archivistCol = "artifact", exact=TRUE)
(objName <- lazyObjectName(hash))
unlink(file.path(tempdir(), "lazyDir"), recursive=TRUE)
}
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyLoad2}}
}

