% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{data}
\name{.lazyREnv}
\alias{.lazyREnv}
\title{The lazyREnv environment}
\format{An object of class \code{environment} of length 0.}
\usage{
.lazyREnv
}
\description{
Environment used internally to store active lazyDir
}
\keyword{datasets}

