% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{lazyRm}
\alias{lazyRm}
\title{Load lazy objects from a \code{lazyR} database}
\usage{
lazyRm(objNames = NULL, lazyDir = "lazyDir", exact = TRUE,
  removeRasterFile = FALSE)
}
\arguments{
\item{objNames}{A character vector of names of objects to be removed from a \code{lazyR} database.}

\item{lazyDir}{Character string of directory to be used for the lazy databases.}
}
\value{
Nothing returned. This function is used for its side effects, i.e., loading lazy objects
}
\description{
Load all named objects into the specified environment.
}
\examples{
\dontrun{
a <- rnorm(10)
lazySave(a, lazyDir=tempdir())
lazyRm("a", lazyDir=tempdir())
}
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyLoad2}}
}

