% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lu}
\alias{lu}
\title{Count unique values in a vector}
\usage{
lu(x)
}
\arguments{
\item{None.}{}
}
\description{
A short cut to typing length(unique()), nothing more.
}
\examples{
lu()
}
\keyword{length}
\keyword{unique}

