% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{make_names_nice}
\alias{make_names_nice}
\title{Clean up data frame column names}
\usage{
make_names_nice(names)
}
\arguments{
\item{None}{}
}
\description{
Remove unsightly characters from character vector. Intended for use on data frame column names.
}
\examples{
names(df) = make_names_nice(names(df))
}
\keyword{column}
\keyword{names}

