% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{month_diff}
\alias{month_diff}
\title{Calculate the number of calender months between dates}
\usage{
month_diff(date1, date2)
}
\arguments{
\item{None.}{}
}
\description{
Author: http://stackoverflow.com/users/143305/dirk-eddelbuettel
}
\examples{
month_diff("1980-03-31", Sys.Date())
}
\keyword{date}
\keyword{month}

