% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{month_year}
\alias{month_year}
\title{Obtain yearmon object as factor for given dates}
\usage{
month_year(x)
}
\arguments{
\item{None}{}
}
\description{
Shortcut for converting to yearmon and then factor (useful for plots).
}
\examples{
month_year(today())
}
\keyword{month}
\keyword{year}

