% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{cols}{Number of columns in layout,
layout A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
Plot multiple ggplot objects on one page.
}
\details{
Author: Winston Chang: http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)

ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)

If the layout is something like matrix(c(1, 2, 3, 3), nrow = 2, byrow = TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\examples{
multiplot(plot1, plot2, cols = 2)
}
\keyword{ggplot2}
\keyword{grid}

