% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helper_functions.R
\docType{data}
\name{nice_chart}
\alias{nice_chart}
\title{Create ggplot object to make chart more pretty}
\format{\preformatted{List of 8
 $ axis.ticks  : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.title  :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 25
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ axis.text   :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 25
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ strip.text  :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 25
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ legend.text :List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 25
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ legend.key  : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ legend.title:List of 8
  ..$ family    : NULL
  ..$ face      : NULL
  ..$ colour    : NULL
  ..$ size      : num 25
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 $ plot.title  :List of 8
  ..$ family    : NULL
  ..$ face      : chr "bold"
  ..$ colour    : NULL
  ..$ size      : num 25
  ..$ hjust     : NULL
  ..$ vjust     : NULL
  ..$ angle     : NULL
  ..$ lineheight: NULL
  ..- attr(*, "class")= chr [1:2] "element_text" "element"
 - attr(*, "class")= chr [1:2] "theme" "gg"
 - attr(*, "complete")= logi FALSE
}}
\usage{
nice_chart
}
\arguments{
\item{None}{}
}
\description{
A custom theme that can be added to ggplot2 objects to make them look less scientific (useful when presenting to non-analysts).
}
\examples{
ggplot(df, aes(x, y)) + geom_point() + nice_chart
}
\keyword{ggplot2}
\keyword{theme}

