% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_resid_stats.R
\name{pca_resid_stats}
\alias{pca_resid_stats}
\title{Examine PCA Residuals}
\usage{
pca_resid_stats(data, loadings)
}
\arguments{
\item{data}{A correlation matrix.}

\item{loadings}{PCA loadings in PCA$loadings format.}
}
\value{
Overview of residual statistics.
}
\description{
Examine PCA Residuals
}
