\name{rearrange2}
\alias{rearrange2}
\title{Interactively rearrange order in data when names are specified.}
\usage{
  rearrange2(data, ..., order = T)
}
\arguments{
  \item{data}{data to rearrange.}

  \item{...}{names of the variable to rearrange}

  \item{order}{boolean indicating if function
  \code{\link{arrange}} should be run as well.}
}
\description{
  Interactively rearrange order in data when names are
  specified.
}

