% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_assume.R
\name{reg_assume}
\alias{reg_assume}
\title{Regression Assumption Plot Grid}
\usage{
reg_assume(model, var)
}
\arguments{
\item{model}{A regression model (lm or lmer)}

\item{var}{The dependent variable used for the regression given in df$var format.}
}
\value{
A four-plot grid summarising the key assumptions of regression.
Plot A shows the trend of the relationship (is it linear?)
Plot B shows the variance
Plot C is a histogram of resituals to look for normality
Plot D is a qq plot to again look at normality of residuals.
}
\description{
Regression Assumption Plot Grid
}
