\name{repmat.i}
\alias{repmat}
\alias{repmat.i}
\alias{repmat.j}
\title{Replicate a matrix}
\usage{
  repmat.i(A, m)

  repmat.j(A, n)

  repmat(A, m, n)
}
\arguments{
  \item{A}{matrix to replicate}

  \item{m}{number of row replication}

  \item{n}{number of colum replication}
}
\description{
  Replicate a matrix
}

