\name{row.names.as.col}
\alias{row.names.as.col}
\title{Add row.names to a data frame as a new column}
\usage{
  row.names.as.col(data, position = 1, var.name)
}
\arguments{
  \item{data}{the data frame}

  \item{position}{position of the new column (1 by
  default)}

  \item{va.name}{name of the newly created column. By
  default \code{var.name=".row"}.}
}
\description{
  Add row.names to a data frame as a new column
}
\seealso{
  \code{\link{col.as.row.names}}
}

