% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{saveToRepoRaster}
\alias{saveToRepoRaster}
\title{Alternative to saveToRepo for rasters}
\usage{
saveToRepoRaster(obj, objName = NULL, lazyDir = NULL, tags = NULL,
  compareRasterFileLength = 1e+06, useRobocopy = TRUE)
}
\arguments{
\item{obj}{The raster object to save to the repository.}

\item{objName}{A character representation of the object name.}

\item{lazyDir}{the lazyDir to use.}

\item{tags}{Optional character vector of tags. Passed to \code{saveToRepo}.}

\item{compareRasterFileLength}{Numeric. This is passed to the length arg in \code{digest}
when determining if the Raster file is already in the database. Default 1e6. Passed to \code{lazySave}.}

\item{useRobocopy}{Logical. For Windows only, this will use a system call to Robocopy which appears to be much
faster than the internal \code{file.copy} function. Uses /MIR flag.}
}
\value{
A raster object and its file backing will be passed to the archivist repository.
}
\description{
Rasters are sometimes file-based, so the normal save mechanism doesn't work. This function creates an
explicit save of the file that is backing the raster, in addition to saving the object metadata in
the archivist repository database.
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyLoad2}}
}

