% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_effect_size.R
\name{t_effect_size}
\alias{t_effect_size}
\title{T-test Effect Size}
\usage{
t_effect_size(test)
}
\arguments{
\item{test}{A t-test saved as an object.}
}
\value{
Effect size (r) for the given test.
}
\description{
T-test Effect Size
}
