% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{week_start}
\alias{week_start}
\title{Return first Monday of week for any given date}
\usage{
week_start(date)
}
\arguments{
\item{None}{}
}
\description{
Basically floor_date but with week start taken to be Monday, rather than Sunday.
}
\examples{
week_start(today())
}
\keyword{floor_date}
\keyword{monday}

