% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{year_quarter}
\alias{year_quarter}
\title{Classify dates according to year and quarter}
\usage{
year_quarter(x)
}
\arguments{
\item{None}{}
}
\description{
Returns a factor vector of same length as input.
}
\examples{
year_quarter(today())
}
\keyword{date}
\keyword{quarter}

