% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_outlier.R
\name{z_outlier}
\alias{z_outlier}
\title{Remove Outliers Based on Z-Scores}
\usage{
z_outlier(var, idvar, condvar, sd, data)
}
\arguments{
\item{var}{\itemize{
\item A continuous variable.
}}

\item{idvar}{\itemize{
\item A participant ID variable.
}}

\item{condvar}{\itemize{
\item A grouping variable indicating condition.
}}

\item{sd}{\itemize{
\item Number of SDs outside of which outliers are defined.
}}

\item{data}{\itemize{
\item A data frame containing the data.
}}
}
\value{
A new data frame trimmed of outliers.
}
\description{
Remove Outliers Based on Z-Scores
}
