% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{copyLazyDir}
\alias{copyLazyDir}
\title{Copy a lazyDir and all the files in it}
\usage{
copyLazyDir(oldLazyDir = NULL, newLazyDir = NULL, overwrite = TRUE,
  copyRasterFile = TRUE, clearRepo = TRUE, create = TRUE,
  silent = FALSE)
}
\arguments{
\item{oldLazyDir}{The source lazyDir}

\item{newLazyDir}{The new lazyDir}

\item{overwrite}{Passed to \code{lazySave}}

\item{copyRasterFile}{Passed to \code{lazySave}}

\item{clearRepo}{Passed to \code{lazySave}}

\item{create}{Passed to \code{checkLazyDir}}

\item{silent}{Should a progress be printed}
}
\description{
Copy a lazyDir and all the files in it
}
\examples{
\dontrun{
obj1 <- 1:10
obj2 <- 11:20
setLazyDir
oldLazyDir <- file.path(tempdir(), "old")
newLazyDir <- file.path(tempdir(), "new")
lazySave(obj1, obj2, lazyDir=oldLazyDir)
copyLazyDir(oldLazyDir, newLazyDir)
}
}
\author{
Eliot McIntire
}

