% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_kable.R
\name{dark_kable}
\alias{dark_kable}
\title{Dark-Theme Friendly Kable}
\usage{
dark_kable(...)
}
\arguments{
\item{...}{see the documentation for knitr::kable}

\item{format}{A character string denoting format. This is set to "html" by default.}

\item{digits}{The number of decimals to print, byd efault this is set to 2.}

\item{table.attr}{By default this is set to "style = \"color:white;\"" which together with format = html gives
white text output in RMarkdown.}
}
\value{
A kable-formated table using html styling to achieve white text.
}
\description{
Dark-Theme Friendly Kable
}
\note{
See ?knitr::kable for more details on each argument. In order to get black text on a white background when
knitting, the following CSS code must be added to the start of the RMarkdown file (and NOT in an R code chunk):\if{html}{\out{
<style>
table \{
 background-color: white !important;
 color: black !important;
\}
</style>
}}
}
