% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{last function}
\usage{
last(v)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
last(y)
}
\keyword{last}

