% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{setLazyDir}
\alias{getLazyDir}
\alias{setLazyDir}
\title{Set and get the lazyR database directory}
\usage{
setLazyDir(lazyDir)

getLazyDir()
}
\arguments{
\item{lazyDir}{A character string to the directory where the lazy database should be kept}
}
\value{
New lazyDir
}
\description{
This can be set and gotten with these functions, or all functions can take
an argument, lazyDir, manually.
}
\examples{
\dontrun{
setLazyDir(file.path(tempdir(), "lazyDir"))
a <- rnorm(10)
lazySave(a)
lazyRm("a")
}
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyLoad2}}
}

