% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{lazyIs}
\alias{lazyIs}
\title{Get object class}
\usage{
lazyIs(objName, class2 = NULL, removeCharacter = TRUE, lazyDir = NULL)
}
\arguments{
\item{objName}{A character string indicating the object name}

\item{removeCharacter}{Logical. There is an artifact tag, class:character for any
non-character object. This will be removed automatically in the return, unless this
is FALSE}
}
\value{
A character vector equivalent to the an \code{is(objName)} command
}
\description{
Get object class
}
\examples{
setLazyDir(file.path(tempdir(), "lazyDir"))
a <- rnorm(10)
lazySave(a)
lazyIs("a", "numeric")
lazyIs("a") # shows all
lazyIs("b") # error, does not exist
lazyRm("a")
unlink(file.path(tempdir(), "lazyDir"))
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyLoad2}}
}

