% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{lazyLoad2}
\alias{lazyLoad2}
\title{Load lazy objects from a \code{lazyR} database}
\usage{
lazyLoad2(objNames = NULL, lazyDir = "lazyDir", envir = parent.frame())
}
\arguments{
\item{objNames}{A character vector of object names to load lazily, usually from a lazyLs call.}

\item{lazyDir}{Character string of directory to be used for the lazy databases.}

\item{envir}{Which environment should the objects be loaded into. Default is \code{.GlobalEnv}}
}
\value{
Nothing returned. This function is used for its side effects, i.e., loading lazy objects
}
\description{
Load all named objects into the specified environment.
}
\examples{
library(SpaDES)
tmpdir <- checkPath(file.path(tempdir(), "lazyDir"), create=TRUE)
obj <- rnorm(10)
# save the obj
lazySave(obj, lazyDir=tmpdir)
# remove the obj
rm(obj)
any(ls()=="obj") # Is FALSE
# load it back in
lazyLoad2("obj", lazyDir=tmpdir)
any(ls()=="obj") # Is TRUE
unlink(tmpdir, recursive = TRUE)
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyRm}}.
}

