% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{lazyObjectName}
\alias{lazyObjectName}
\title{Get object name with an md5hash}
\usage{
lazyObjectName(md5Hash, lazyDir)
}
\arguments{
\item{md5Hash}{A character string indicating the md5hash value to look up the object name}
}
\value{
An object name
}
\description{
Get object name with an md5hash
}
\examples{
\dontrun{
}
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyLoad2}}
}

