% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR-package.R
\docType{package}
\name{lazyR-package}
\alias{lazyR}
\alias{lazyR-package}
\title{Categorized overview of the \code{lazyR} package}
\description{
Bug reports: \url{https://github.com/PredictiveEcology/lazyR/issues}
}
\section{1. lazy load databases}{

}

\section{1.1 Saving}{



\tabular{ll}{
  \code{\link{lazySave}} \tab save objects to a lazy database\cr
}
}

\section{1.2 Loading}{


\tabular{ll}{
  \code{\link{lazyLoad2}} \tab load lazy databases\cr
}
}
\author{
Eliot J. B. McIntire \email{Eliot.McIntire@canada.ca}
}
\keyword{------------------------------------------------------------------------------------------}
\keyword{package}

