% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{data}
\name{.lazyREnv}
\alias{.lazyREnv}
\title{The lazyREnv environment}
\format{\preformatted{<environment: 0x00000000175cea58> 
}}
\usage{
.lazyREnv
}
\description{
Environment used internally to store active lazyDir
}
\keyword{datasets}

