% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{lazyRm}
\alias{lazyRm}
\title{Load lazy objects from a \code{lazyR} database}
\usage{
lazyRm(objNames = NULL, lazyDir = NULL, exact = TRUE,
  removeRasterFile = FALSE)
}
\arguments{
\item{objNames}{A character vector of names of objects to be removed from a \code{lazyR} database.}

\item{lazyDir}{Character string of directory to be used for the lazy databases.}

\item{exact}{Should the character string matching be exact, i.e., ^objNames$ in regexp.}

\item{removeRasterFile}{Logical. If true then the source raster file will be removed in addition to the
database entry. This is only true if the file is in the lazyDir}
}
\value{
Nothing returned. This function is used for its side effects, i.e., loading lazy objects
}
\description{
Load all named objects into the specified environment.
}
\examples{
\dontrun{
a <- rnorm(10)
lazySave(a, lazyDir=tempdir())
lazyRm("a", lazyDir=tempdir())
}
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyLoad2}}
}

