% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lazyR.R
\docType{methods}
\name{lazySave}
\alias{lazySave}
\title{Save objects to a \code{lazyR} database}
\usage{
lazySave(..., lazyDir = "lazyDir", tags = NULL, clearRepo = FALSE,
  overwrite = FALSE)
}
\arguments{
\item{...}{Objects to save to an R lazy database, with metadata saved via archivist package repository.}

\item{lazyDir}{Character string of directory to be used for the lazy databases. This creates an archivist repository.}

\item{tags}{Character vector of additional tags to add to the objects being saved}

\item{clearRepo}{Logical. Should the repository be deleted before adding new objects}

\item{overwrite}{Logical. Should new object replace (i.e,. overwrite) previous object with same name}
}
\value{
invisibly returns a character vector of objects saved.
}
\description{
This function creates an archivist repository that holds metadata ("tags") associated
with the object. Unlike the archivist package, the objects themselves are saved as
lazy load databases (i.e., via tools:::makeLazyLoadDB). Each object is saved into its own
lazy load database, with a filename that matches the object name. Because it is also an
archivist repository, lazy loading of objects can be done via tags, classes etc..
}
\details{
The main use case for this is large, complicated datasets, such as GIS databases, that
are used within R. Loading them lazily means that the user can have access to all of them,
including their characteristics, without loading them all first (could be very slow).
}
\examples{
a <- rnorm(10)
b <- rnorm(20)
lazySave(a, b)
\dontrun{
lazySave(mget(ls()))
}
}
\author{
Eliot McIntire
}
\seealso{
\code{\link{lazyLs}}, \code{\link{lazyRm}}, \code{\link{lazyLoad2}}.
}

